Fix line line break insertion
